Improve css loading and format
Moved css to separate file only for the weekly-news section so that it wouldn't be loaded (wasting bandwidth) on pages that don't need it. Did basic formatting on new css file and fixed some most basic css problems like duplicate properties; combined padding; and removing font-family from  this level, as it should be defined on a higher level for the entire theme in the theme customizer.