"Small" refactor 😋
There were a few problems with scaling, so I changed CSS to a more hierarchically style, removed unnecessary css properties. Also made the section more configurable and simplified the amount calculation to use a (configurable) tag. NB! The calculation is still not entirely correct and can be wrong especially for collections with >1000 items, as this is the absolute maximum that Shopify allows us to iterate :(