Jesse dev (#5)
* Add show_image_label setting

* Add show_image_label setting as false

* Add div styling for image_label

* Add logic for creating image-label & add setting for show_image_label

* Add show_image_label setting as false

* Remove custom-block

* Add translatable string, change logic in main-product, change styling px -> rem

* Hotfix

* main-product changes

* Image label border colour opacity change

* Add styling

* Main page for weekly-news

* Styling changes

* Rename section

* Remove not needed section file

* Improve css loading and format

Moved css to separate file only for the weekly-news section so that it wouldn't be loaded (wasting bandwidth) on pages that don't need it. Did basic formatting on new css file and fixed some most basic css problems like duplicate properties; combined padding; and removing font-family from  this level, as it should be defined on a higher level for the entire theme in the theme customizer.

* Improve liquid/html formatting

Indent everything properly and move some elements around for better readability

* Improve liquid/html formatting

Indent everything properly and move some elements around for better readability

* "Small" refactor 😋

There were a few problems with scaling, so I changed CSS to a more hierarchically style, removed unnecessary css properties. Also made the section more configurable and simplified the amount calculation to use a (configurable) tag. NB! The calculation is still not entirely correct and can be wrong especially for collections with >1000 items, as this is the absolute maximum that Shopify allows us to iterate :(

Co-authored-by: Jesse Hyvärinen <jessehyvarinen@Kameratori-MacBook-Air-2.local>
Co-authored-by: Karl Rankla <karlrank@gmail.com>